Created MusicLanesTweener + prefab
MusicLanesTweener is the object that controls the animations of the lanes. This can be used to notify the player of successful/unsuccessful button presses - currently it just has the one animation for each lane. 

Also created prefab for this object - when dragging into the working scene, please link up the lane game objects for this to work. 

Main scene also requires the art to be modified  - can use assets from the Art Implementation Test scene.